fix some earlier printf mistakes

+++ src/pmm.h
@@ -6,7 +6,7 @@
 #include "const.h"
 
 void pmm_init(multiboot_info_t *info, u32 magic, u32 bitmap_location);
-void *pmm_allocate_blocks(u64 size);
-void pmm_free_blocks(void *address, u64 size);
+void *pmm_allocate_blocks(u32 size);
+void pmm_free_blocks(void *address, u32 size);
 
 #endif
